build external packages with vite instead of rollup
