+++ package.json
@@ -82,7 +82,7 @@
 	},
 	"scripts": {
 		"dev": "vite",
-		"build": "tsc --project ./src/tsconfig.json && rollup -c ./src/rollup.config.js",
+		"build": "tsc --project ./src/tsconfig.json && vite build -c vite.cms.config.ts",
 		"build:vite": "tsc && vite build --mode staging",
 		"build:for:static": "vite build",
 		"build:for:cms": "npm run build && node ./devops/build/copy-to-cms.js",
@@ -134,9 +134,6 @@
 		"@mdx-js/react": "^2.2.1",
 		"@open-wc/testing": "^3.1.7",
 		"@playwright/test": "^1.30.0",
-		"@rollup/plugin-commonjs": "^25.0.0",
-		"@rollup/plugin-json": "^6.0.0",
-		"@rollup/plugin-node-resolve": "^15.0.1",
 		"@storybook/addon-a11y": "7.0.20",
 		"@storybook/addon-actions": "7.0.20",
 		"@storybook/addon-essentials": "7.0.20",
@@ -173,19 +170,13 @@
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
 		"remark-gfm": "^3.0.1",
-		"rollup": "^3.21.2",
-		"rollup-plugin-dts": "^5.2.0",
-		"rollup-plugin-esbuild": "^5.0.0",
-		"rollup-plugin-import-css": "^3.2.1",
-		"rollup-plugin-url": "^3.0.1",
-		"rollup-plugin-web-worker-loader": "^1.6.1",
 		"storybook": "7.0.20",
 		"tiny-glob": "^0.2.9",
 		"tsc-alias": "^1.8.6",
 		"typescript": "^5.1.3",
 		"typescript-json-schema": "^0.56.0",
-		"vite": "^4.3.3",
-		"vite-plugin-static-copy": "^0.15.0",
+		"vite": "^4.3.9",
+		"vite-plugin-static-copy": "^0.16.0",
 		"vite-tsconfig-paths": "^4.2.0",
 		"web-component-analyzer": "^2.0.0-next.4"
 	},
