+++ src/external/uui/index.ts
@@ -1,1 +1,3 @@
+import '@umbraco-ui/uui-css/dist/uui-css.css';
+
 export * from '@umbraco-ui/uui';
